+++ java-io-base/src/main/java/internal/io/FunctionalFileParser.java
@@ -8,6 +8,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Objects;
 
+import static nbbrd.io.Resource.uncloseableInputStream;
+
 @lombok.RequiredArgsConstructor
 public final class FunctionalFileParser<T> implements FileParser<T> {
 
@@ -15,6 +17,6 @@ public final class FunctionalFileParser<T> implements FileParser<T> {
 
     @Override
     public @NonNull T parseStream(@NonNull InputStream resource) throws IOException {
-        return Objects.requireNonNull(function.applyWithIO(new UncloseableInputStream(resource)), "result");
+        return Objects.requireNonNull(function.applyWithIO(uncloseableInputStream(resource)), "result");
     }
 }