+++ java-io-base/src/main/java/module-info.java
@@ -28,6 +28,5 @@ module nbbrd.io.base {
     exports nbbrd.io.text;
     exports nbbrd.io.zip;
 
-    exports internal.io to nbbrd.io.xml, nbbrd.io.xml.bind, nbbrd.io.picocsv;
     exports internal.io.text to nbbrd.io.xml, nbbrd.io.xml.bind, nbbrd.io.picocsv;
 }