+++ java-io-xml/src/main/java/nbbrd/io/xml/Sax.java
@@ -16,7 +16,6 @@
  */
 package nbbrd.io.xml;
 
-import internal.io.UncloseableInputStream;
 import internal.io.text.LegacyFiles;
 import lombok.NonNull;
 import nbbrd.design.StaticFactoryMethod;
@@ -35,6 +34,8 @@ import java.util.Locale;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
+
+import static nbbrd.io.Resource.uncloseableInputStream;
 
 /**
  * @author Philippe Charles
@@ -164,12 +165,12 @@ public class Sax {
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource) throws IOException {
-            return doParse(new InputSource(new UncloseableInputStream(resource)));
+            return doParse(new InputSource(uncloseableInputStream(resource)));
         }
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource, @NonNull Charset encoding) throws IOException {
-            InputSource input = new InputSource(new UncloseableInputStream(resource));
+            InputSource input = new InputSource(uncloseableInputStream(resource));
             input.setEncoding(encoding.name());
             return doParse(input);
         }